+++ src/core/errors/not-found-error.ts
@@ -1,0 +1,10 @@
+import CustomError from "./custom-error";
+
+class NotFoundError extends CustomError {
+    constructor(message?: string) {
+        super(message || "Not found");
+        this.name = "NotFound";
+        Object.setPrototypeOf(this, NotFoundError.prototype);
+    }
+}
+export default NotFoundError;