+++ src/core/ports/user-repository.port.ts
@@ -1,7 +1,8 @@
+import { RegisterInfo } from "../domain/user/register-info";
 import { User } from "../domain/user/user";
 
 interface UserRepository {
-    persist(user: Omit<User, "id">): Promise<null | User>;
+    persist(user: RegisterInfo): Promise<null | User>;
     findByEmail(email: string): Promise<null | User>;
 }
 export default UserRepository;