+++ src/infrastructure/logger/get-logger.ts
@@ -1,10 +1,10 @@
 import winston from "./winston-logger";
 
 const getLogger = () => {
-    const info = (message: String) => {
+    const info = (message: string) => {
         return winston.info(message);
     };
-    const error = (message: String) => {
+    const error = (message: string) => {
         return winston.error(message);
     };
     return {