+++ src/interface/server/app.ts
@@ -1,4 +1,4 @@
-import express, { Application } from "express";
+import express, { Application, Request, Response, NextFunction } from "express";
 import registerController from "./controllers/register.controller";
 import cors from "cors";
 import swaggerUi from "swagger-ui-express";
@@ -6,7 +6,9 @@ import swaggerOptions from "./api-docs/swagger-options";
 import { container } from "../../infrastructure/dependency-injection/awilix-set-up";
 import { scopePerRequest } from "awilix-express";
 import loginController from "./controllers/login.controller";
-import CustomError from "../../infrastructure/errors/custom-error";
+import registerAdminController from "./controllers/admin/register.admin.controller";
+import validateAdmin from "./validate-admin";
+import CustomError from "../../core/errors/custom-error";
 
 export const createServer = (port: number) => {
     const app: Application = express();
@@ -14,9 +16,6 @@ export const createServer = (port: number) => {
     app.use(express.json());
     app.use(express.urlencoded({ extended: true }));
     app.use(scopePerRequest(container));
-    app.use((err: any, req: any, res: any, next: any) => {
-        res.status(500).send(err);
-    });
 
     app.use("/api-docs", swaggerUi.serve, swaggerUi.setup(swaggerOptions));
 
@@ -24,8 +23,13 @@ export const createServer = (port: number) => {
 
     app.post("/register", registerController); // makeInvoker(registerController));
 
-    app.use((err: any, req: any, res: any, next: any) => {
-        res.status(500).send({ message: err.message });
+    app.use("/admin", validateAdmin);
+    app.post("/admin/register", registerAdminController);
+
+    app.use((error: any, req: Request, res: Response, next: NextFunction) => {
+        res.status(error.statusCode).send({
+            message: error.responseBody,
+        });
     });
 
     return {