+++ src/interface/server/controllers/login.controller.test.ts
@@ -27,7 +27,7 @@ describe("loginController", () => {
         const next = jest.fn;
         const sendResponse = jest.fn();
         const res: any = {
-            status: jest.fn((x) => ({
+            status: jest.fn(() => ({
                 send: sendResponse,
             })),
         };
@@ -36,7 +36,7 @@ describe("loginController", () => {
         expect(mockValidateLogin).toHaveBeenCalledWith(req.body);
         expect(res.status).toHaveBeenCalledWith(200);
         expect(sendResponse).toHaveBeenCalledWith({
-            message: { token: req.body },
+            token: req.body,
         });
     });
 });