+++ CHANGELOG.md
@@ -39,3 +39,29 @@ straightforward as possible. Here we would have the update steps for 0.2.0 for p
 ### Fixed
 
 -   [FIXED REGISTER USE CASE TEST PARTY](https://www.notion.so/Register-Use-Case-Test-party-5f2d2123e0bf43f6bb4a2f0fc7d6ea5d)
+
+## [0.2.1] - 2022-05-12
+
+Here we write upgrading notes for brands. It's a team effort to make them as
+straightforward as possible. Here we would have the update steps for 0.2.1 for people to follow.
+
+### Added
+
+-   [CREATE ADMIN REGISTER ENDPOINT](https://github.com/DomingoCast/jnt-copy/pull/153)
+
+-   [CREATE ERROR HANDLER AND HTTP ERROR HANDLER](https://github.com/DomingoCast/jnt-copy/pull/158)
+
+-   [CREATE ROL IN USER](https://github.com/DomingoCast/jnt-copy/pull/152)
+
+
+### Changed
+
+-   [CHANGE OF LOCATION OF THE ERROR FOLDER](https://github.com/DomingoCast/jnt-copy/pull/165)
+
+-   [CHANGED STATUS CODE OF SOME HTTP ERROR](https://github.com/DomingoCast/jnt-copy/pull/158)
+
+-   [CHANGED TOKEN RESPONSE](https://github.com/DomingoCast/jnt-copy/pull/171)
+
+### Fixed
+
+-   [FIXED SOME FAILS INTO LOGIN TEST PARTY](https://github.com/DomingoCast/jnt-copy/pull/158)